Bug#21317507: Generated columns: stored column rejected, but virtual is accepted
Verification patch: Validate that virtual and stored columns are
generated consistently.

Currently, this test is only enabled for MyISAM, InnoDB has the problem
reported in the bug report: Adding a virtual column does not cause
the underlying expression to be validated for consistency with the
data type specified for the generated column.