BUG#21127387 - TRANSACTION-WRITE-SET-EXTRACTION CONSIDERS NON UNIQUE AND NULL KEYS.
Problem : Transaction write set extraction feature of MySQL Replication
is incorrectly considering non unique keys and unique with NULL values.
This is because the code which iterates through the keys in the table to
form the write set value does not have any mechanism to skip the non
unique keys and the unique keys which can take up NULL value. This is
causing problem with other replication modules like Group replication
which ends up rolling back some transactions which it should not.

Fix : The problem is fixed by adding a check for the non unique keys and
NULL valued unique keys.