Bug#21324507 INNODB: ASSERT(DICT_TF_GET_REC_FORMAT(TABLE->FLAGS) != REC_FORMAT_COMPRESSED
This patch fixes a mistake in handling the incompatibility between TEMPORARY
and DATA DIRECTORY.  When innodb_strict_mode is OFF, DATA DIRECTORY should be
ignored.  If strict mode is ON, no assert occurs because the CREATE TABLE is rejected.
if strict mode is OFF the assert occurs when creating the table because
instead of unsetting the HAS_DATA_DIR flag, the code turns off the
FILE_PER_TABLE flag.