Bug#20444737 STRING::CHOP ASSERTS ON NAUGHTY TABLE NAMES
DESCRIPTION:

This is a regression from the fix of Bug#16066637.
Bug#16066637 introduced a character set conversion while writing
to binlog. While calculating the length of the converted string,
system default charset was being used instead of the new charset.
This led to assert in debug build and crash in asan release build.

FIX :

Use the new charset to which string is converted to calculate the
length of the string.