Bug#21286221: DEBUG SERVER CRASHES ON RESET SLAVE WITH               LOCKED VERSION TOKEN
The problem was that RESET/FLUSH statements could trigger a debug
assert if the current connection had acquired any locking
service locks. This had no consequences for release builds.

This patch fixes the problem by adjusting the assert so
that it takes the possibility of locking service locks
into account - similar to what it already did for user level
locks.