Bug#21682439: BACKPORT OF BUG#20928289 TO MYSQL 5.6
Backport from mysql-5.7 to mysql-5.6 of:

Bug#20928289: DEFAULT CMAKE EXCLUDES UTF8_5624_1 CHARSET,
              REQUIRED FOR MAIN.CTYPE_LDML

The problem was that on platforms where char is unsigned,
the server was unable to parse collation defintions which
included non-7bit ASCII characters. Affected platforms
include ARM and PowerPC.

The reason was that the parsing code in once place used
'> 0' to check if a character was 7 bit or not. This
will fail if the char is unsigned.

This patch fixes the problem by rewriting the check to
check the MSB explicitly.

No test added as coverage is already included by
existing tests - e.g. main.ctype_ldml.

This patch is a modified version of the contributed patch.

(cherry picked from commit d59ac89547a75e1a422ac4e806fe7b853489764d)