Bug#21341482: IS_SERVER_ENGINE_READY() METHOD IS NO LONGER USED
The is_server_engine_ready() was a method, in the server, that could
be used by plugins to know that such structures like the slave
repository info were initialized.
The method however, being based on the tc_log variable, no longer
works and is also redundant due to the channel service interface.
With the channel interface, the developer in the plugin can simply
verify if the default channel exists.