Bug#21355630: MYSQL_CONFIG_EDITOR ASSUMES SIGNED CHAR
The problem was that check_and_create_login_file() inside
mysql_config_editor.cc is declared to return my_bool but
instead returns 0/-1. Since my_bool is implemented as char,
this becomes a problem for platforms where char is unsigned
such as ARM and PowerPC. When the return value is assigned
to int, -1 is no longer -1.

The net consequence is that in some cases mysql_config_editor
can report success when the operation has actually failed.

This patch fixes the problem by using TRUE/FALSE (=1/0)
instead of -1/0.