WL#6631: Detect transaction boundaries
Adds a new sysvar "session_track_transaction_info" which
defaults to "OFF".

When set to "STATE", transaction state will be reported
to the client (not in a transaction, in implicit trx, in
explicit trx, trx reads attached, unsafe reads attached,
trx writes attached, unsafe writes attached, LOCK TABLES
used, etc.). This helps load balancers and the like make
decisions about whether it is safe to switch a given
connection away from a server (using just the "in
transaction" flag is useless for long BEGIN ... COMMIT
AND CHAIN ... COMMIT AND CHAIN ... well, chains).

When set to "CHARACTERISTICS", information of how to
restore transaction characteristics (WITH CONSISTENT
SNAPSHOT, READ ONLY|WRITE, ISOLATION LEVEL, etc.) is
reported in addition to the above "STATE".

To subscribe to this feature, a client needs to be linked
against a libmysql recent enough to support tracker items /
extended OK messages.