Bug#21068487 - INNODB: PREVENT RELATIVE ISL PATHS UNDER THE DATADIR
Post-fix Valgrind error.  A test to make sure that the separator is at the end of
fil_path_to_mysql_datadir was checking the wrong byte, one byte past what
was initialized so far.  This did not cause a problem because this conditional check
is did not occur in our MTR tests.  It was only noticed in valgrind.

(cherry picked from commit 20fb0b52771f0f333650ee4f734758605dfaa3e3)