BUG#20106837: ALTER TABLE WHICH DROPS AND ADDS THE SAME FULLTEXT               INDEX IS NOT INPLACE/FAST.
Analysis
-------

The ALTER TABLE operation which drops and adds the same FULLTEXT
index is not a INPLACE/FAST operation.

While determining if the key definition has changed, the index
algorithm is compared between the old index and new index
definition. While preparing the new index list during the ALTER
TABLE operation, the FULLTEXT index algorithm is not set correctly.
Hence the ALTER TABLE operation for dropping and adding the
same FULLTEXT index is not a fast operation.

Also, such fast operation which drops and adds the same index
did not clean up the temporary 'frm' created during the
operation.

Fix
---
a) Set the algorithm correctly for the FULLTEXT index in the
   new index list prepared.
b) Delete the temporary 'frm' file created for such fast/INPLACE
   operations.