Bug#21381806: JSON: ASSERTION FAILED: ARG->NULL_VALUE
The failing assertion asserts that null_value is true if the field
type of an item is MYSQL_TYPE_NULL. This happens before a val_*
function has been called on the item, and null_value is not
necessarily updated until one of the val_* functions is called.

This patch calls update_null_value() on the item to make sure that
null_value is up to date before we try to use it.