BUG#21368299 - SERVER CRASH AT SHUTDOWN AFTER PORT CLASH DETECTED.
MySQLD server crashes when the server is started with the port
number of an existing server.

Due to port clash, initialization of the acceptor object fails
and the instantiated acceptor object is deleted. During cleanup
on shutdown, the acceptor object is again freed. This results
in corruption of the heap and thereby resulting in a crash.

The fix is to free the acceptor object as part of cleanup during
shutdown. Also reset all listener objects to NULL after free and
close the socket listener if the listener is not closed when the
socket listener is undergoing destruction.

(cherry picked from commit 0a6ae5839ee3bb8af24c511c3a63d9c200b67638)