Bug#21277074: crash (segfault) in THD::change_item_tree on exec of prep
The problem here is when resolving a field during execution of a
prepared statement, when the field is an aggregation argument,
an outer reference, and comes from a derived table or a view, and
the aggregation function is aggregated on the outer level.

When this type of field is resolved for the second time, some code
that updates max_arg_level of the aggregation function is triggered.
In this case, max_arg_level was set to the level according to the
current select_lex, but should be set to the level which the table
the field belongs to is resolved from.

For evidence, check the other assignments to max_arg_level inside
Item_field::fix_fields() and Item_field::fix_outer_field() and
Item_ref:fix_fields().