Bug #21280801: VERSION TOKEN LOCKING DOES NOT WORK
The plugin was releasing all the get_lock() locks at the end of
each query regardless of whether session version tokens are present
or not.
Fixed as suggested by Ingo by making the lock release dependent on
presence of session version tokens.
Added some more verbose comments on what the function is doing.