Bug #20728894: MEMORY LEAK IN ADD_DERIVED_KEY()
Problem:
Possible derived keys are created in stmt_arena's mem_root which
is the permanent memory area for statement. It does not get deleted
after execution of the query. In case of stored procedure, if a query
creating derived table is executed many times, the memory keeps growing.

Solution:
Allocated derived keys in thd->mem_root which is for runtime objects instead
of thd->stmt_arena->mem_root.