Bug#21391781 ASSERT WHEN RUNNING ALTER TABLE ON A TABLE WITH INDEX ON              VIRTUAL COLUMN
During ALTER TABLE on a table that has an indexed virtual column, the
storage engine calls my_eval_gcolumn_expr() to compute the value of
the virtual column. Information about which columns that should be evaluated
is specified using an ulonglong used a bit map. This bitmap can only
define indexes on the 64 first columns.

In the failing test, the virtual column with an index was column
number 2. In addition column number 66 was a virtual column. Due to a
bug in how the code checked if a column should be evaluated or not,
column 66 left-shifted 66 times in the bitmap matched the same
position as column number 2 in the bitmap. This caused column 66 to be
evaluated. But since the storage engine had not requested this column
to be evaluated, it had not provided the values for the base columns
needed for this evaluation. This caused the evaluation to be based on
whatever data was in the record buffer for the base columns and in
some cases this could cause errors and hitting the assert.

The fix for this problem is to replace the ulonglong used as a 64 bit
bitmap with a general bitmap of type MY_BITMAP that can have a bit for
all columns in the table.

This change also removes the restriction that for virtual columns,
only the first 64 columns in a table can be indexed.