Bug#21437989: ASSERTION FAILED: JSON_BINARY::PARSE_BINARY(PTR, LENGTH)               .IS_VALID()
Field_json::store_binary() has an assertion which checks that the
supplied binary string is a valid binary representation of a JSON
document. In some cases the binary string is not a valid
representation of a JSON document. Instead it is an empty string, and
this causes the assertion to fail.

This happens if the value is the result of an INSERT statement that
successfully inserted an empty value into a not nullable JSON column.
Such insertions are usually not allowed and cause an error, but the
errors could be suppressed and downgraded to warnings if INSERT IGNORE
is used or strict SQL mode is not in effect.

It is not those insertions that trigger the assert, as the insertions
of the empty values don't go through Field_json. The assert is fired
when an empty JSON value, already in a table, is attempted copied into
another JSON column.

This patch relaxes the assertion so that it doesn't fire when trying
to store an empty binary string. Field_json::val_json() already has
special handling for zero-length binary strings, in which case it
returns the JSON null literal, so allowing such values to be stored
should not cause any problems when reading them out again.