Bug#21448719: WRONG RESULT FOR JSON VALUE IN OUTER JOIN WITH VIEW
Item_direct_view_ref overrides all the val_*() functions in order to
ensure that they return NULL for the null-extended values of the inner
table of an outer join. It does not override val_json(), so JSON
values could in some cases come out as not NULL when NULL is expected.

This patch provides an override for val_json() in
Item_direct_view_ref, so that JSON gets the same handling of NULL as
the other data types.