Bug#21335821: --INITIALIZE OPTION DON'T ACCEPT "INSTALL PLUGIN" STATEMENTS IN INIT-FILE
The --skip-grant-tables mode (implicit with --initialize/--bootstrap)
was preventing the INSTALL PLUGIN and UNINSTALL PLUGIN commands.
This was not necessary, since the plugins are initialized regardless of
--skip-grant-tables and the only thing missing would be the ACL tables
so these commands can check access to the plugin system table.

Fixed by removing the blanket conditions for these two commands and
just don't checking access to the plugin system table when ACLs don't
exist.

Test case added.