BUG#20582189 - INNODB.DOUBLEWRITE FAILS DUE TO DATABASE CORRUPTION
Follow-up patch for this bug.

After previous fixes, current failure is due to the design of the test case.
This test case could not make sure that the page it modifes when the server
is down has a correct copy in doublewrite buffer. The test case always flush
the the to be modified page first, and then start a transaction to update
some other pages. It's possible that pages changed by the transaction would
occupy the same doublewrite buffer page with the one used by the to be
modifed page, so server may reports the corrupted page. We can simply always
flush the to be modified page at last to fix this issue.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>
rb: 9270