Bug#21397107 ASSERTION FAILED: K < STATIC_CAST<SIZE_T>(GET_DIMENSION()) IN GIS_POINT::GET<0>(
Issue:
When passing geometry data to GIS functions, the user is allowed
to send raw byte data for some of the functions:

  SELECT ST_AsText(0x00000000010100000000000000000024400000000000002440);

However, if the user add extra garbage bytes at the end of a valid byte
input, an assertion is hit in some of the GIS functions.

Fix:
Check that there are no extra trailing bytes in GEOMETRY byte strings
manually crafted as input to geometry functions. If the check fails, take the
input as invalid.