BUG#21357008 SLAVE_OPEN_TEMP_TABLES WRONG AFTER RESET SLAVE WHEN USING MULTI-SOURCE
Problem:
RESET SLAVE FOR CHANNEL 'channel' forced the status variable
Slave_open_temp_tables value to 0, even if there was another
replication channel with open temporary tables which were
not affected by the statement.
In addition, the function that modifies the value for the
variable did not consider that multiple threads could update
it even when MTS was not enabled.
Finally, read operations to the variable were not safe in a
multiple-writers multiple-readers scenario.

Fix:
The variable slave_open_temp_tables is now an atomic variable,
so all concurrent accesses are safe. The variable is now only
decremented the number of tables that were actually closed.
Introduced another atomic variable to keep track of the open
temporary tables on a per channel basis, so that closing
or changing the master of a given replication channel will
issue a warning only if there are temporary tables open in
that channel.
Finally, since the warning does not convoy any information
regarding open temp tables in channels, when the slave is
stopped and there are open temp tables in more that one
channel, only one warning is issued.