Bug#21273010 SLAVE GROUP EVENT PARSER DOES NOT RECOGNIZE XA ROLLBACK AS A BOUNDARY
Problem:
A warning was issued when a XA transaction rollbacked without
a preceeding XA PREPARE command when binlog format was set
to statement and binlog_direct_non_transactional_updates is
set to off.
XA ROLLBACK was an EVENT_BOUNDARY_TYPE_STATEMENT, which
would put the parser in the EVENT_PARSER_DML state. Since
is it followed by an EVENT_BOUNDARY_TYPE_GTID, this event
sequence was not valid.

Fix:
Added a new boundary_type so that the parser will be the
correct state (EVENT_PARSER_NONE) after a XA ROLLBACK command.