Fixed bug#21469535: VALGRIND ERROR (CONDITIONAL JUMP) WHEN INSERT ROWS INTO A PARTITIONED TABLE
When prune partitions, the server only cares which partitions it needs scan.
During this time, only partition expression needs an evaluation. So there is
an extra argument named 'bitmap' to show which columns needs an evaluation in
fill_record.

However, update_generated_write_fields tries to update the virtual generated
columns only if they are in table->write_set and ignore the 'bitmap' in
fill_record. Because the base columns of vitual generated columns in
table->write_set don't occur in 'bitmap', it results in an valgrind error
during generate the value of virtual generated columns.

The solution is to add the 'bitmap' argument to update_generated_write_fields.
If the field index of virtual generated column doesn't occur in 'bitmap'
argument, the virtual generated column won't be evaluated.