Bug#21487833: DBUG_ABORT() IN JSON_WRAPPER::MAKE_HASH_KEY WITH ORDERED JSON
If a NULL value is inserted into a not nullable JSON column (which is
possible if strict SQL mode is turned off), an empty string is
inserted at that position. Since an empty string is not a valid
representation of a JSON value, code that attempts to parse it, might
raise an error, or possibly hit an assertion in debug builds.

Field_json::val_json() knows about this case and will return a
representation of the JSON null literal if it encounters it. However,
some parts of the code bypass val_json() and access the internal
representation directly, and they fail when they try to parse the
stored value.

This patch fixes the failures by making the failing code access the
JSON values via Field_json::val_json(), which doesn't suffer from this
problem, instead of accessing the internal representation directly.