Bug#21216067 ASSERTION FAILED ROW_UPD_SEC_INDEX_ENTRY (INNOBASE/ROW/ROW0UPD.CC:2103)
A multi-table UPDATE updates a table using a temporary table:
- it first runs a regular join (as a SELECT), storing in the
tmp table the new (future) value of columns and the rowid of the row
- then it scans the tmp table and for each row:
  * fetches the row by rowid into record[1]
  * copy new values from tmp table to record[0]
  * calls before-update trigger
  * calls update_row()
but it forgets to calculate the new value of generated columns.
This is specific of UPDATE-with-tmp-table (Query_result_update::do_updates());
when on-the-fly UPDATE is used Query_result_update::send_data()
calls fill_record() which does the job.
Fix: call update_generated_write_fields().
This solves all wrong results and the crash in the report.